pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quillan"
version = "1.2.0"
description = "Filesystem-driven AI story generation engine"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.30",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
tokens = ["tiktoken>=0.7"]
cover = ["Pillow>=10.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff",
    "mypy",
    "types-aiofiles",
    "types-python-jose",
]

[project.scripts]
quillan = "quillan.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["quillan*"]

[tool.setuptools.package-data]
quillan = ["templates/*.tex", "templates/*.txt", "model_pricing.yaml", "py.typed"]

[tool.setuptools.data-files]
"." = ["quillan.env.example"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
